Moved routes classes under Routes sub-namespace
