+++ src/RouteCompiler.php
@@ -17,7 +17,7 @@ declare(strict_types=1);
 
 namespace Flight\Routing;
 
-use Flight\Routing\{GeneratedUri, Route};
+use Flight\Routing\Routes\{FastRoute as Route, Route as BaseRoute};
 use Flight\Routing\Exceptions\{UriHandlerException, UrlGenerationException};
 use Flight\Routing\Interfaces\RouteCompilerInterface;
